pyproject.toml
[tool.poetry]
name = "meisengeige-bot"
version = "0.1.0"
description = "Daily monitoring bot for Meisengeige events at Cinecitta with Telegram notifications"
authors = ["Grigorii Dorotynskii"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.14"
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.1.0"
python-telegram-bot = "^21.0"
python-dotenv = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
ruff = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py314']

[tool.ruff]
line-length = 100
target-version = "py314"
